+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 4.5.0
+
+* Adds support to set whether to draw the scrollbar. See
+  `AndroidWebViewController.setVerticalScrollBarEnabled`,
+  `AndroidWebViewController.setHorizontalScrollBarEnabled`,
+  `AndroidWebViewController.supportsSetScrollBarsEnabled`.
+
 ## 4.4.2
 
 * Updates pigeon generated code to fix `ImplicitSamInstance` and `SyntheticAccessor` Kotlin lint